+++ src/main/java/com/travel/role/domain/room/dto/TravelEssentialReqDTO.java
@@ -1,6 +1,6 @@
 package com.travel.role.domain.room.dto;
 
-import static com.travel.role.global.exception.ExceptionMessage.*;
+import static com.travel.role.global.exception.dto.ExceptionMessage.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
