+++ src/main/java/com/travel/role/domain/user/service/UserService.java
@@ -1,6 +1,6 @@
 package com.travel.role.domain.user.service;
 
-import static com.travel.role.global.exception.ExceptionMessage.*;
+import static com.travel.role.global.exception.dto.ExceptionMessage.*;
 
 import java.time.LocalDate;
 import java.util.Objects;
