+++ src/main/java/com/travel/role/global/auth/service/mail/MailService.java
@@ -1,6 +1,6 @@
 package com.travel.role.global.auth.service.mail;
 
-import static com.travel.role.global.exception.ExceptionMessage.*;
+import static com.travel.role.global.exception.dto.ExceptionMessage.*;
 
 import javax.mail.SendFailedException;
 import javax.mail.internet.MimeMessage;
@@ -8,9 +8,6 @@ import javax.mail.internet.MimeMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
-
-import com.travel.role.global.auth.service.AuthService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
