+++ src/main/java/com/travel/role/global/exception/common/ResourceOperationAccessDeniedException.java
@@ -1,6 +1,6 @@
 package com.travel.role.global.exception.common;
 
-import com.travel.role.global.exception.ExceptionMessage;
+import com.travel.role.global.exception.dto.ExceptionMessage;
 
 public class ResourceOperationAccessDeniedException extends RuntimeException {
