+++ src/main/java/com/travel/role/global/exception/common/S3ImageNotFoundException.java
@@ -1,6 +1,6 @@
 package com.travel.role.global.exception.common;
 
-import com.travel.role.global.exception.ExceptionMessage;
+import com.travel.role.global.exception.dto.ExceptionMessage;
 
 public class S3ImageNotFoundException extends RuntimeException {
