+++ src/main/java/com/travel/role/global/exception/room/AlreadyExistInRoomException.java
@@ -1,6 +1,6 @@
 package com.travel.role.global.exception.room;
 
-import static com.travel.role.global.exception.ExceptionMessage.*;
+import static com.travel.role.global.exception.dto.ExceptionMessage.*;
 
 public class AlreadyExistInRoomException extends RuntimeException{
 	public AlreadyExistInRoomException() {super(ALREADY_EXIST_USER_IN_ROOM);}
